ARM: KVM: KVM_DEV_TYPE_ARM_VGIC_V2 is changed to 5
In the latest migration/vgic code, KVM_DEV_TYPE_ARM_VGIC_V2 has been changed,
because 4 is occupied.

This code is not so good, we should also sync the kvm source code to the latest
version. But the latest version has few problems with our dirty page tracking..
We need to use older version in this stage, just for temporality.